# Conflicts:
#	features/user/loginregister/build.gradle
#	features/user/loginregister/src/main/java/com/tokopedia/loginregister/login/di/LoginModule.java
#	features/user/loginregister/src/main/java/com/tokopedia/loginregister/login/view/fragment/LoginEmailPhoneFragment.kt

+++ features/travel/hotel/src/main/java/com/tokopedia/hotel/hoteldetail/presentation/fragment/HotelReviewFragment.kt
@@ -48,7 +48,7 @@ class HotelReviewFragment : BaseListFragment<HotelReview, ReviewAdapterTypeFacto
         super.onCreate(savedInstanceState)
 
         arguments?.let {
-            param.propertyId = it.getInt(ARG_PROPERTY_ID, 0)
+            param.propertyId = it.getLong(ARG_PROPERTY_ID, 0)
         }
 
         val viewModelProvider = ViewModelProviders.of(this, viewModelFactory)
@@ -183,10 +183,10 @@ class HotelReviewFragment : BaseListFragment<HotelReview, ReviewAdapterTypeFacto
         const val COUNTRY_ID = "id"
         const val COUNTRY_ALL = "all"
 
-        fun createInstance(propertyId: Int): HotelReviewFragment {
+        fun createInstance(propertyId: Long): HotelReviewFragment {
             return HotelReviewFragment().also {
                 it.arguments = Bundle().apply {
-                    putInt(ARG_PROPERTY_ID, propertyId)
+                    putLong(ARG_PROPERTY_ID, propertyId)
                 }
             }
         }
